requirements.txt
streamlit
pandas
numpy
plotly
python-pptx
